Delay creating symlinks until after all regular files
When decompressing a tarball there isn't any guarantee of the order
which tar headers are presented to us by (*tar.Reader).Next(), this
means that it's possible for us to evaluate and try to create a symlink
before the file which it points to has been created.

This is usually fine, a symlink isn't required to point to anything.
However, for security reasons (mitigating against "Zip-slip") we wan't
to evaluate that a symlink resolves to a file that is within the
expected directory (i.e. it should not escape outside of the directory
which we are decompressing into).

For this reason, we need to collect all symlinks and process them after
we've gone through all regular files.

There might still be a chance of problems being caused by symlinks that
point to other symlinks -- I'll cross that bridge if I get to it.