Setup ping.Queue endpoint
Allow enqueueing within the ping service.

Moves some stuff around regarding database pool initialisation.
Adds and Enqueuer, Dequeuer, and Queue interface to the queue package,
which do what they say on the tin.

Changes the signature of Enq and Deq to be more explicit and rely less
or state within the struct.