Allow setting --commit on slugcmplr release
`slugcmplr` can be invoked on a separate (but equal) commit to the one
that is eventually released. This can lead to pushing commit IDs which
do not match any commit ID on the default branch for a project.

This can occur especially in scenarios where slugs are pre-emptively
built on auto generated merge-commits between a branch and its target.

This change allows for a manual override of the commit associated with a
release.